moved upload functionality to dialog
